% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_ate.R
\name{coef.online_ate}
\alias{coef.online_ate}
\title{Extract the current ATE/ATT estimate from an `online_ate` object}
\usage{
\method{coef}{online_ate}(object, ...)
}
\arguments{
\item{object}{An `online_ate` object.}

\item{...}{Ignored.}
}
\value{
A named scalar: the current point estimate.
}
\description{
Extract the current ATE/ATT estimate from an `online_ate` object
}
