% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_super_learner.R
\name{coef.online_sl}
\alias{coef.online_sl}
\title{Extract current ensemble weights from an `online_sl` object}
\usage{
\method{coef}{online_sl}(object, ...)
}
\arguments{
\item{object}{An `online_sl` object.}

\item{...}{Ignored.}
}
\value{
Named numeric vector of current ensemble weights.
}
\description{
Extract current ensemble weights from an `online_sl` object
}
