\name{coef.stream_fit}
\alias{coef.stream_fit}
\title{Extract coefficients from a stream_fit object}
\usage{
\method{coef}{stream_fit}(object, ...)
}
\arguments{
\item{object}{A \code{stream_fit} object.}
\item{...}{Ignored.}
}
\value{
Named numeric vector of final coefficients.
}
\description{
Extract method for objects of class \code{stream_fit}.
}
