% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_learner.R
\name{coef.super_learner_fit}
\alias{coef.super_learner_fit}
\title{Extract meta-learner weights from a `super_learner_fit` object}
\usage{
\method{coef}{super_learner_fit}(object, ...)
}
\arguments{
\item{object}{A `super_learner_fit` object.}

\item{...}{Ignored.}
}
\value{
Named numeric vector of meta-learner weights.
}
\description{
Extract meta-learner weights from a `super_learner_fit` object
}
