% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isgd.R
\name{.implicit_fn}
\alias{.implicit_fn}
\title{Implicit equation for iSGD (internal)}
\usage{
.implicit_fn(xi, y, eta, gamma, norm_x2, linkinv)
}
\description{
Implicit equation for iSGD (internal)
}
\keyword{internal}
