% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_learner.R
\name{make_learner}
\alias{make_learner}
\title{Construct a learner specification}
\usage{
make_learner(fit, predict, name = NULL)
}
\arguments{
\item{fit}{A function with signature `function(X, y, family, ...)` that
returns a fitted model object.}

\item{predict}{A function with signature `function(model, newdata, ...)` that
returns a numeric vector of **response-scale** predictions.}

\item{name}{Optional character string used as a label. If `NULL`, the
learner is labelled by its position in the library.}
}
\value{
A list of class `sl_learner` with elements `fit`, `predict`,
  and `name`.
}
\description{
Creates a learner object for use in [super_learner_fit()]. A learner wraps
a `fit` function and a `predict` function behind a common interface.
}
\examples{
\donttest{
lrn <- make_learner(
  fit     = function(X, y, family, ...) rls_glm_fit(X, y, family = family),
  predict = function(model, newdata, ...) predict(model, newdata)
)
}

}
\seealso{
[super_learner_fit()]
}
