% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_ate.R
\name{print.online_ate}
\alias{print.online_ate}
\title{Print an `online_ate` object}
\usage{
\method{print}{online_ate}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An `online_ate` object.}

\item{digits}{Number of significant digits. Default `4`.}

\item{...}{Ignored.}
}
\description{
Print an `online_ate` object
}
