% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_super_learner.R
\name{print.online_sl}
\alias{print.online_sl}
\title{Print an `online_sl` object}
\usage{
\method{print}{online_sl}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An `online_sl` object.}

\item{digits}{Number of significant digits. Default `4`.}

\item{...}{Ignored.}
}
\description{
Print an `online_sl` object
}
