\name{print.stream_fit}
\alias{print.stream_fit}
\title{Print a stream_fit object}
\usage{
\method{print}{stream_fit}(x, digits = 4L, ...)
}
\arguments{
\item{x}{A \code{stream_fit} object.}
\item{digits}{Number of significant digits. Default \code{4}.}
\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for objects of class \code{stream_fit}.
}
