% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_learner.R
\name{print.super_learner_fit}
\alias{print.super_learner_fit}
\title{Print a `super_learner_fit` object}
\usage{
\method{print}{super_learner_fit}(x, digits = 4L, ...)
}
\arguments{
\item{x}{A `super_learner_fit` object.}

\item{digits}{Number of significant digits. Default `4`.}

\item{...}{Ignored.}
}
\description{
Print a `super_learner_fit` object
}
