\name{summary.stream_fit}
\alias{summary.stream_fit}
\title{Summary of a stream_fit object}
\usage{
\method{summary}{stream_fit}(object, ...)
}
\arguments{
\item{object}{A \code{stream_fit} object.}
\item{...}{Ignored.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for objects of class \code{stream_fit}. Currently delegates
to \code{\link{print.stream_fit}}.
}
